Logic centralization
Creates a canvasmanager class that interacts with both the usecanvas and canvas components and their implementation.
